requirements.txt
pycairo==1.19.1
numpy==1.19.0
progress==1.5.0